narrow guilds_foods_changes.total to SMALLINT UNSIGNED #4220

+++ Core/src/core/database/logs/migrations/021-fix-guilds-foods-changes-total.ts
@@ -4,13 +4,13 @@ import {
 
 /**
  * Widens `guilds_foods_changes.total` from TINYINT UNSIGNED (max 255) to
- * MEDIUMINT UNSIGNED (max 16_777_215). A guild's accumulated food storage
- * can largely exceed 255, which made the original column overflow and crash
- * the insert at runtime.
+ * SMALLINT UNSIGNED (max 65_535). A guild's food storage snapshot can
+ * exceed 255 (current max pantry cap is 1000 for common food), which made
+ * the original column overflow and crash the insert at runtime.
  */
 export async function up({ context }: { context: QueryInterface }): Promise<void> {
 	await context.changeColumn("guilds_foods_changes", "total", {
-		type: DataTypes.MEDIUMINT.UNSIGNED,
+		type: DataTypes.SMALLINT.UNSIGNED,
 		allowNull: false
 	});
 }
